+++ GitSccProvider/SccProviderService.DocumentEvents.cs
@@ -14,13 +14,14 @@ using ThreadHelper = Microsoft.VisualStudio.Shell.ThreadHelper;
 
 namespace GitScc
 {
-    public partial class SccProviderService
+    public partial class SccProviderService : IVsSolutionLoadEvents
     {
         // The cookie for project document events
         private uint _tpdTrackProjectDocumentsCookie;
        // private DTE2 _activeIde;
         private WindowEvents _windowEvents;
         private SolutionEvents _solutionEvents;
+        private bool _solutionLoaded = false;
 
         private void SetupDocumentEvents()
         {
@@ -34,7 +35,7 @@ namespace GitScc
             //activeIde.Events.SolutionItemsEvents.
             _windowEvents = activeIde.Events.WindowEvents;
             _solutionEvents = activeIde.Events.SolutionEvents;
-            _solutionEvents.Opened += _solutionEvents_Opened;
+            _solutionEvents.ProjectAdded += _solutionEvents_ProjectAdded;
             _windowEvents.WindowActivated += _windowEvents_WindowActivated;
 
         }
@@ -49,14 +50,48 @@ namespace GitScc
                 _tpdTrackProjectDocumentsCookie = VSConstants.VSCOOKIE_NIL;
             }
             _windowEvents.WindowActivated -= _windowEvents_WindowActivated;
-            _solutionEvents.Opened -= _solutionEvents_Opened;
+            _solutionEvents.ProjectAdded -= _solutionEvents_ProjectAdded;
+            _windowEvents.WindowActivated -= _windowEvents_WindowActivated;
         }
+
+        private async void _solutionEvents_ProjectAdded(Project dteProject)
+        {
+            try
+            {
+               await AutoAddProject(dteProject);
+            }
+            catch (Exception)
+            {
+                //TODO 
+            }
+        }
+
+        private async Task AutoAddProject(Project dteProject)
+        {
+            await ThreadHelper.JoinableTaskFactory.SwitchToMainThreadAsync();
+            var project = await SolutionExtensions.GetIVsHierarchy(dteProject) as IVsSccProject2;
+            if (_solutionLoaded) 
+            {
+                if (GitSccOptions.Current.AutoAddProjects && project != null && !SolutionExtensions.IsProjectInGit(dteProject.FullName))
+                {
+                    await AddProjectToSourceControl(project);
+                }
+                //ok To be safe rebuild cache 
+                await RefreshSolution();
+                //if (!_fileCache.ProjectAddedToCache(project))
+                //{
+                //    await _fileCache.AddProject(project);
+                //}
+            }
+           
+        }
+
+
 
         private async void _solutionEvents_Opened()
         {
             await _statusBarManager.SetActiveRepository(RepositoryManager.Instance.SolutionTracker);
-            await EnableSccForSolution();
-            await ReloadAllGlyphs();
+            await RefreshSolution();
         }
 
         private async void _windowEvents_WindowActivated(Window GotFocus, Window LostFocus)
@@ -128,7 +163,47 @@ namespace GitScc
         public int OnAfterAddFilesEx(int cProjects, int cFiles, IVsProject[] rgpProjects, int[] rgFirstIndices,
             string[] rgpszMkDocuments, VSADDFILEFLAGS[] rgFlags)
         {
-            return VSConstants.E_NOTIMPL;
+
+            if (Active)
+            {
+                // Start by iterating through all projects calling this function
+                for (int iProject = 0; iProject < cProjects; iProject++)
+                {
+                    IVsSccProject2 sccProject = rgpProjects[iProject] as IVsSccProject2;
+                    IVsHierarchy pHier = rgpProjects[iProject] as IVsHierarchy;
+
+                    // If the project is not controllable, or is not controlled, skip it
+                    if (sccProject == null)
+                    {
+                        continue;
+                    }
+
+                    // Files in this project are in rgszMkOldNames, rgszMkNewNames arrays starting with iProjectFilesStart index and ending at iNextProjecFilesStart-1
+                    int iProjectFilesStart = rgFirstIndices[iProject];
+                    int iNextProjecFilesStart = cFiles;
+                    if (iProject < cProjects - 1)
+                    {
+                        iNextProjecFilesStart = rgFirstIndices[iProject + 1];
+                    }
+
+                    // Now that we know which files belong to this project, iterate the project files
+                    for (int iFile = iProjectFilesStart; iFile < iNextProjecFilesStart; iFile++)
+                    {
+                        var fileName = rgpszMkDocuments[iFile];
+                        _fileCache.AddFile(fileName,sccProject);
+                        if (GitSccOptions.Current.AutoAddFiles)
+                        {
+                            var repo = RepositoryManager.Instance.GetTrackerForPath(fileName);
+                            repo.AddFile(fileName);
+                        }
+                        // Refresh the solution explorer glyphs for all projects containing this file
+                        //IList<VSITEMSELECTION> nodes = GetControlledProjectsContainingFile(rgpszMkDocuments[iFile]);
+                        //_sccProvider.RefreshNodesGlyphs(nodes);
+                    }
+                }
+            }
+
+            return VSConstants.E_NOTIMPL; ;
         }
 
         public int OnAfterAddDirectoriesEx(int cProjects, int cDirectories, IVsProject[] rgpProjects, int[] rgFirstIndices,
@@ -201,5 +276,44 @@ namespace GitScc
         }
 
         #endregion
+
+        public int OnBeforeOpenSolution(string pszSolutionFilename)
+        {
+            _solutionLoaded = false;
+            return VSConstants.S_OK;
+        }
+
+        public int OnBeforeBackgroundSolutionLoadBegins()
+        {
+            return VSConstants.S_OK;
+        }
+
+        public int OnQueryBackgroundLoadProjectBatch(out bool pfShouldDelayLoadToNextIdle)
+        {
+            pfShouldDelayLoadToNextIdle = false;
+            return VSConstants.S_OK;
+        }
+
+        public int OnBeforeLoadProjectBatch(bool fIsBackgroundIdleBatch)
+        {
+            return VSConstants.S_OK;
+        }
+
+        public int OnAfterLoadProjectBatch(bool fIsBackgroundIdleBatch)
+        {
+            return VSConstants.S_OK;
+        }
+
+        public int OnAfterBackgroundSolutionLoadComplete()
+        {
+            _solutionLoaded = true;
+
+            ThreadHelper.JoinableTaskFactory.Run(async delegate
+            {
+                await RefreshSolution();
+            });
+
+            return VSConstants.S_OK;
+        }
     }
 }